Bar 93 federate lasso net

+++ src/fl/datasets/memory.py
@@ -50,7 +50,7 @@ def load_from_pgen(pfile_path: str, gwas_path: str, snp_count: int, sample_indic
 
 def load_phenotype(phenotype_path: str) -> numpy.ndarray:
     data = pandas.read_table(phenotype_path)
-    return data.iloc[:, -1].values
+    return data.iloc[:, -1].values.astype(numpy.float32)
 
 
 def load_covariates(covariates_path: str, load_pcs: bool = False) -> numpy.ndarray:
